requirements.txt
apache-airflow==2.7.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
scikit-learn==1.3.1
dvc==3.30.1  # Changed from python-dvc to dvc
dvc[gdrive]  # Add this for Google Drive support

